Yoni - fixed some
