+++ src/main/java/frc/robot/JoysticksBindings.java
@@ -1,9 +1,7 @@
 package frc.robot;
 
-import edu.wpi.first.wpilibj2.command.PrintCommand;
 import frc.robot.constants.Ports;
 import frc.utils.controllers.joysticks.SmartJoystick;
-import frc.utils.controllers.keyboard.KeyboardController;
 
 public class JoysticksBindings {
 
@@ -15,15 +13,12 @@ public class JoysticksBindings {
 
     private static final SmartJoystick FOURTH_JOYSTICK = new SmartJoystick(Ports.JoystickDriverStationPorts.FOURTH);
 
-    private static final KeyboardController KEYBOARD_CONTROLLER = new KeyboardController();
-
 
     public static void configureBindings() {
         mainJoystickButtons();
         secondJoystickButtons();
         thirdJoystickButtons();
         fourthJoystickButtons();
-        keyboardButtons();
     }
 
     private static void mainJoystickButtons() {
@@ -45,13 +40,4 @@ public class JoysticksBindings {
         SmartJoystick usedJoystick = FOURTH_JOYSTICK;
         // bindings
     }
-
-    private static void keyboardButtons() {
-        KeyboardController usedJoystick = KEYBOARD_CONTROLLER;
-        usedJoystick.A.onTrue(new PrintCommand("A PRESSED"));
-        usedJoystick.A.onFalse(new PrintCommand("A RELEASED"));
-        usedJoystick.B.onTrue(new PrintCommand("B PRESSED"));
-        usedJoystick.B.onFalse(new PrintCommand("B RELEASED"));
-    }
-
 }
